requirements.txt
asgiref==3.8.1
certifi==2024.2.2
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.1
dnspython==2.6.1
filelock==3.14.0
mongoengine==0.28.2
pipenv==2023.12.1
platformdirs==4.2.1
pymongo==4.7.2
setuptools==69.5.1
sqlparse==0.5.0
virtualenv==20.26.1
